+++ src/resources/zero-trust/devices/posture/posture.ts
@@ -1,6 +1,7 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import { APIResource } from '../../../../resource';
+import { isRequestOptions } from '../../../../core';
 import * as Core from '../../../../core';
 import * as IntegrationsAPI from './integrations';
 import {
@@ -36,7 +37,7 @@ export class Posture extends APIResource {
     params: PostureCreateParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<DevicePostureRule | null> {
-    const { account_id, ...body } = params;
+    const { account_id = this._client.accountId, ...body } = params;
     return (
       this._client.post(`/accounts/${account_id}/devices/posture`, { body, ...options }) as Core.APIPromise<{
         result: DevicePostureRule | null;
@@ -65,7 +66,7 @@ export class Posture extends APIResource {
     params: PostureUpdateParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<DevicePostureRule | null> {
-    const { account_id, ...body } = params;
+    const { account_id = this._client.accountId, ...body } = params;
     return (
       this._client.put(`/accounts/${account_id}/devices/posture/${ruleId}`, {
         body,
@@ -88,10 +89,18 @@ export class Posture extends APIResource {
    * ```
    */
   list(
-    params: PostureListParams,
+    params?: PostureListParams,
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<DevicePostureRulesSinglePage, DevicePostureRule>;
+  list(options?: Core.RequestOptions): Core.PagePromise<DevicePostureRulesSinglePage, DevicePostureRule>;
+  list(
+    params: PostureListParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.PagePromise<DevicePostureRulesSinglePage, DevicePostureRule> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.list({}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return this._client.getAPIList(
       `/accounts/${account_id}/devices/posture`,
       DevicePostureRulesSinglePage,
@@ -113,10 +122,19 @@ export class Posture extends APIResource {
    */
   delete(
     ruleId: string,
-    params: PostureDeleteParams,
+    params?: PostureDeleteParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<PostureDeleteResponse | null>;
+  delete(ruleId: string, options?: Core.RequestOptions): Core.APIPromise<PostureDeleteResponse | null>;
+  delete(
+    ruleId: string,
+    params: PostureDeleteParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<PostureDeleteResponse | null> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.delete(ruleId, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return (
       this._client.delete(`/accounts/${account_id}/devices/posture/${ruleId}`, options) as Core.APIPromise<{
         result: PostureDeleteResponse | null;
@@ -138,10 +156,19 @@ export class Posture extends APIResource {
    */
   get(
     ruleId: string,
-    params: PostureGetParams,
+    params?: PostureGetParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<DevicePostureRule | null>;
+  get(ruleId: string, options?: Core.RequestOptions): Core.APIPromise<DevicePostureRule | null>;
+  get(
+    ruleId: string,
+    params: PostureGetParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<DevicePostureRule | null> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.get(ruleId, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return (
       this._client.get(`/accounts/${account_id}/devices/posture/${ruleId}`, options) as Core.APIPromise<{
         result: DevicePostureRule | null;
@@ -1162,7 +1189,7 @@ export interface PostureCreateParams {
   /**
    * Path param
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Body param: The name of the device posture rule.
@@ -1229,7 +1256,7 @@ export interface PostureUpdateParams {
   /**
    * Path param
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Body param: The name of the device posture rule.
@@ -1293,15 +1320,15 @@ export interface PostureUpdateParams {
 }
 
 export interface PostureListParams {
-  account_id: string;
+  account_id?: string;
 }
 
 export interface PostureDeleteParams {
-  account_id: string;
+  account_id?: string;
 }
 
 export interface PostureGetParams {
-  account_id: string;
+  account_id?: string;
 }
 
 Posture.DevicePostureRulesSinglePage = DevicePostureRulesSinglePage;
