+++ src/resources/zero-trust/dlp/entries/entries.ts
@@ -1,6 +1,7 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import { APIResource } from '../../../../resource';
+import { isRequestOptions } from '../../../../core';
 import * as Core from '../../../../core';
 import * as CustomAPI from './custom';
 import {
@@ -69,7 +70,7 @@ export class Entries extends APIResource {
    * ```
    */
   create(params: EntryCreateParams, options?: Core.RequestOptions): Core.APIPromise<EntryCreateResponse> {
-    const { account_id, ...body } = params;
+    const { account_id = this._client.accountId, ...body } = params;
     return (
       this._client.post(`/accounts/${account_id}/dlp/entries`, { body, ...options }) as Core.APIPromise<{
         result: EntryCreateResponse;
@@ -98,7 +99,7 @@ export class Entries extends APIResource {
     params: EntryUpdateParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<EntryUpdateResponse> {
-    const { account_id, ...body } = params;
+    const { account_id = this._client.accountId, ...body } = params;
     return (
       this._client.put(`/accounts/${account_id}/dlp/entries/${entryId}`, {
         body,
@@ -121,10 +122,18 @@ export class Entries extends APIResource {
    * ```
    */
   list(
-    params: EntryListParams,
+    params?: EntryListParams,
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<EntryListResponsesSinglePage, EntryListResponse>;
+  list(options?: Core.RequestOptions): Core.PagePromise<EntryListResponsesSinglePage, EntryListResponse>;
+  list(
+    params: EntryListParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.PagePromise<EntryListResponsesSinglePage, EntryListResponse> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.list({}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return this._client.getAPIList(
       `/accounts/${account_id}/dlp/entries`,
       EntryListResponsesSinglePage,
@@ -145,10 +154,19 @@ export class Entries extends APIResource {
    */
   delete(
     entryId: string,
-    params: EntryDeleteParams,
+    params?: EntryDeleteParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<EntryDeleteResponse | null>;
+  delete(entryId: string, options?: Core.RequestOptions): Core.APIPromise<EntryDeleteResponse | null>;
+  delete(
+    entryId: string,
+    params: EntryDeleteParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<EntryDeleteResponse | null> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.delete(entryId, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return (
       this._client.delete(`/accounts/${account_id}/dlp/entries/${entryId}`, options) as Core.APIPromise<{
         result: EntryDeleteResponse | null;
@@ -169,10 +187,19 @@ export class Entries extends APIResource {
    */
   get(
     entryId: string,
-    params: EntryGetParams,
+    params?: EntryGetParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<EntryGetResponse>;
+  get(entryId: string, options?: Core.RequestOptions): Core.APIPromise<EntryGetResponse>;
+  get(
+    entryId: string,
+    params: EntryGetParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<EntryGetResponse> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.get(entryId, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return (
       this._client.get(`/accounts/${account_id}/dlp/entries/${entryId}`, options) as Core.APIPromise<{
         result: EntryGetResponse;
@@ -757,7 +784,7 @@ export interface EntryCreateParams {
   /**
    * Path param
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Body param
@@ -795,7 +822,7 @@ export declare namespace EntryUpdateParams {
     /**
      * Path param
      */
-    account_id: string;
+    account_id?: string;
 
     /**
      * Body param
@@ -827,7 +854,7 @@ export declare namespace EntryUpdateParams {
     /**
      * Path param
      */
-    account_id: string;
+    account_id?: string;
 
     /**
      * Body param
@@ -844,7 +871,7 @@ export declare namespace EntryUpdateParams {
     /**
      * Path param
      */
-    account_id: string;
+    account_id?: string;
 
     /**
      * Body param
@@ -859,15 +886,15 @@ export declare namespace EntryUpdateParams {
 }
 
 export interface EntryListParams {
-  account_id: string;
+  account_id?: string;
 }
 
 export interface EntryDeleteParams {
-  account_id: string;
+  account_id?: string;
 }
 
 export interface EntryGetParams {
-  account_id: string;
+  account_id?: string;
 }
 
 Entries.EntryListResponsesSinglePage = EntryListResponsesSinglePage;
